+++ dbldatagen/data_analyzer.py
@@ -8,7 +8,7 @@ This module defines the ``DataAnalyzer`` class.
 This code is experimental and both APIs and code generated is liable to change in future versions.
 """
 from pyspark.sql.types import LongType, FloatType, IntegerType, StringType, DoubleType, BooleanType, ShortType, \
-    TimestampType, DateType, DecimalType, ByteType, BinaryType,  StructType, ArrayType, DataType
+    TimestampType, DateType, DecimalType, ByteType, BinaryType, StructType, ArrayType, DataType
 
 import pyspark.sql as ssql
 import pyspark.sql.functions as F
@@ -21,8 +21,8 @@ class DataAnalyzer:
     """This class is used to analyze an existing data set to assist in generating a test data set with similar
     characteristics, and to generate code from existing schemas and data
 
-    :param df: Spark data frame to analyze
-    :param sparkSession: spark session instance to use when performing spark operations
+    :param df: Spark dataframe to analyze
+    :param sparkSession: Spark session instance to use when performing spark operations
 
     .. warning::
        Experimental
@@ -45,7 +45,7 @@ class DataAnalyzer:
 
     def __init__(self, df=None, sparkSession=None):
         """ Constructor:
-        :param df: Data frame to analyze
+        :param df: Dataframe to analyze
         :param sparkSession: Spark session to use
         """
         assert df is not None, "dataframe must be supplied"
@@ -71,12 +71,12 @@ class DataAnalyzer:
                              dfSummary=None):
         """ Add a measure to the summary dataframe
 
-        :param measureName: name of measure
-        :param summaryExpr: summary expression
+        :param measureName: Name of measure
+        :param summaryExpr: Summary expression
         :param fieldExprs: list of field expressions (or generator)
-        :param dfData: source data df - data being summarized
-        :param rowLimit: number of rows to get for measure
-        :param dfSummary: summary df
+        :param dfData: Source data df - data being summarized
+        :param rowLimit: Number of rows to get for measure
+        :param dfSummary: Summary df
         :return: dfSummary with new measure added
         """
         assert dfData is not None, "source data dataframe must be supplied"
@@ -98,7 +98,7 @@ class DataAnalyzer:
     def summarizeToDF(self):
         """ Generate summary analysis of data set as dataframe
 
-        :return: summary results as dataframe
+        :return: Summary results as dataframe
 
         The resulting dataframe can be displayed with the ``display`` function in a notebook environment
         or with the ``show`` method.
@@ -160,7 +160,7 @@ class DataAnalyzer:
         for row in describeData:
             measure = row['summary']
 
-            values = { k[0]: '' for k in dtypes}
+            values = {k[0]: '' for k in dtypes}
 
             row_key_pairs = row.asDict()
             for k1 in row_key_pairs:
@@ -177,8 +177,8 @@ class DataAnalyzer:
     def summarize(self, suppressOutput=False):
         """ Generate summary analysis of data set and return / print summary results
 
-        :param suppressOutput:  if False, prints results to console also
-        :return: summary results as string
+        :param suppressOutput:  If False, prints results to console also
+        :return: Summary results as string
         """
         dfSummary = self.summarizeToDF()
 
@@ -198,13 +198,35 @@ class DataAnalyzer:
         return summary
 
     @classmethod
-    def _generatorDefaultAttributesFromType(cls, sqlType, dataSummary=None, sourceDf=None):
+    def _valueFromSummary(cls, dataSummary, colName, measure, defaultValue):
+        """ Get value from data summary
+
+        :param dataSummary: Data summary to search, optional
+        :param colName: Column name of column to get value for
+        :param measure: Measure name of measure to get value for
+        :param defaultValue: Default value if any other argument is not specified or value could not be found in
+                             data summary
+        :return: Value from lookup or `defaultValue` if not found
+        """
+        if dataSummary is not None and colName is not None and measure is not None:
+            if measure in dataSummary:
+                measureValues = dataSummary[measure]
+
+                if colName in measureValues:
+                    return measureValues[colName]
+
+        # return default value if value could not be looked up or found
+        return defaultValue
+
+    @classmethod
+    def _generatorDefaultAttributesFromType(cls, sqlType, colName=None, dataSummary=None, sourceDf=None):
         """ Generate default set of attributes for each data type
 
-        :param sqlType: instance of `pyspark.sql.types.DataType`
-        :param dataSummary: map of maps of attributes from data summary, optional
-        :param sourceDf: source dataframe to retrieve attributes of real data, optional
-        :return: attribute string for supplied sqlType
+        :param sqlType: Instance of `pyspark.sql.types.DataType`
+        :param colName: Name of column being generated
+        :param dataSummary: Map of maps of attributes from data summary, optional
+        :param sourceDf: Source dataframe to retrieve attributes of real data, optional
+        :return: Attribute string for supplied sqlType
 
         When generating code from a schema, we have no data heuristics to determine how data should be generated,
         so goal is to just generate code that produces some data.
@@ -216,19 +238,29 @@ class DataAnalyzer:
         if sqlType == StringType():
             result = """template=r'\\\\w'"""
         elif sqlType in [IntegerType(), LongType()]:
-            result = """minValue=1, maxValue=1000000"""
+            minValue = cls._valueFromSummary(dataSummary, colName, "min", defaultValue=0)
+            maxValue = cls._valueFromSummary(dataSummary, colName, "max", defaultValue=1000000)
+            result = f"""minValue={minValue}, maxValue={maxValue}"""
         elif sqlType == ByteType():
-            result = """minValue=1, maxValue=127"""
+            minValue = cls._valueFromSummary(dataSummary, colName, "min", defaultValue=0)
+            maxValue = cls._valueFromSummary(dataSummary, colName, "max", defaultValue=127)
+            result = f"""minValue={minValue}, maxValue={maxValue}"""
         elif sqlType == ShortType():
-            result = """minValue=1, maxValue=32767"""
+            minValue = cls._valueFromSummary(dataSummary, colName, "min", defaultValue=0)
+            maxValue = cls._valueFromSummary(dataSummary, colName, "max", defaultValue=32767)
+            result = f"""minValue={minValue}, maxValue={maxValue}"""
         elif sqlType == BooleanType():
             result = """expr='id % 2 = 1'"""
         elif sqlType == DateType():
             result = """expr='current_date()'"""
         elif isinstance(sqlType, DecimalType):
-            result = """minValue=1, maxValue=1000"""
+            minValue = cls._valueFromSummary(dataSummary, colName, "min", defaultValue=0)
+            maxValue = cls._valueFromSummary(dataSummary, colName, "max", defaultValue=1000)
+            result = f"""minValue={minValue}, maxValue={maxValue}"""
         elif sqlType in [FloatType(), DoubleType()]:
-            result = """minValue=1.0, maxValue=1000000.0, step=0.1"""
+            minValue = cls._valueFromSummary(dataSummary, colName, "min", defaultValue=0.0)
+            maxValue = cls._valueFromSummary(dataSummary, colName, "max", defaultValue=1000000.0)
+            result = f"""minValue={minValue}, maxValue={maxValue}, step=0.1"""
         elif sqlType == TimestampType():
             result = """begin="2020-01-01 01:00:00", end="2020-12-31 23:59:00", interval="1 minute" """
         elif sqlType == BinaryType():
@@ -238,9 +270,9 @@ class DataAnalyzer:
         return result
 
     @classmethod
-    def scriptDataGeneratorFromSchema(cls, schema, suppressOutput=False, name=None):
+    def _scriptDataGeneratorCode(cls, schema, dataSummary=None, sourceDf=None, suppressOutput=False, name=None):
         """
-        Generate outline data generator code from an existing data frame
+        Generate outline data generator code from an existing dataframe
 
         This will generate a data generator spec from an existing dataframe. The resulting code
         can be used to generate a data generation specification.
@@ -249,10 +281,12 @@ class DataAnalyzer:
         For most uses, it will require further modification - however it provides a starting point
         for generation of the specification for a given data set.
 
-        The data frame to be analyzed is the data frame passed to the constructor of the DataAnalyzer object.
+        The dataframe to be analyzed is the dataframe passed to the constructor of the DataAnalyzer object.
 
         :param schema: Pyspark schema - i.e manually constructed StructType or return value from `dataframe.schema`
-        :param suppressOutput: suppress printing of generated code if True
+        :param dataSummary: Map of maps of attributes from data summary, optional
+        :param sourceDf: Source dataframe to retrieve attributes of real data, optional
+        :param suppressOutput: Suppress printing of generated code if True
         :param name: Optional name for data generator
         :return: String containing skeleton code
 
@@ -287,12 +321,15 @@ class DataAnalyzer:
 
             if isinstance(fld.dataType, ArrayType):
                 col_type = fld.dataType.elementType.simpleString()
-                field_attributes = cls._generatorDefaultAttributesFromType(fld.dataType.elementType)
+                field_attributes = cls._generatorDefaultAttributesFromType(fld.dataType.elementType)  # no data look up
                 array_attributes = """structType='array', numFeatures=(2,6)"""
                 name_and_type = f"""'{col_name}', '{col_type}'"""
                 stmts.append(indent + f""".withColumn({name_and_type}, {field_attributes}, {array_attributes})""")
             else:
-                field_attributes = cls._generatorDefaultAttributesFromType(fld.dataType)
+                field_attributes = cls._generatorDefaultAttributesFromType(fld.dataType,
+                                                                           colName=col_name,
+                                                                           dataSummary=dataSummary,
+                                                                           sourceDf=sourceDf)
                 stmts.append(indent + f""".withColumn('{col_name}', '{col_type}', {field_attributes})""")
         stmts.append(indent + ")")
 
@@ -302,9 +339,33 @@ class DataAnalyzer:
 
         return "\n".join(stmts)
 
+    @classmethod
+    def scriptDataGeneratorFromSchema(cls, schema, suppressOutput=False, name=None):
+        """
+        Generate outline data generator code from an existing dataframe
+
+        This will generate a data generator spec from an existing dataframe. The resulting code
+        can be used to generate a data generation specification.
+
+        Note at this point in time, the code generated is stub code only.
+        For most uses, it will require further modification - however it provides a starting point
+        for generation of the specification for a given data set.
+
+        The dataframe to be analyzed is the dataframe passed to the constructor of the DataAnalyzer object.
+
+        :param schema: Pyspark schema - i.e manually constructed StructType or return value from `dataframe.schema`
+        :param suppressOutput: Suppress printing of generated code if True
+        :param name: Optional name for data generator
+        :return: String containing skeleton code
+
+        """
+        return cls._scriptDataGeneratorCode(schema,
+                                            suppressOutput=suppressOutput,
+                                            name=name)
+
     def scriptDataGeneratorFromData(self, suppressOutput=False, name=None):
         """
-        Generate outline data generator code from an existing data frame
+        Generate outline data generator code from an existing dataframe
 
         This will generate a data generator spec from an existing dataframe. The resulting code
         can be used to generate a data generation specification.
@@ -313,9 +374,9 @@ class DataAnalyzer:
         For most uses, it will require further modification - however it provides a starting point
         for generation of the specification for a given data set
 
-        The data frame to be analyzed is the data frame passed to the constructor of the DataAnalyzer object
+        The dataframe to be analyzed is the Spark dataframe passed to the constructor of the DataAnalyzer object
 
-        :param suppressOutput: suppress printing of generated code if True
+        :param suppressOutput: Suppress printing of generated code if True
         :param name: Optional name for data generator
         :return: String containing skeleton code
 
@@ -323,4 +384,16 @@ class DataAnalyzer:
         assert self._df is not None
         assert type(self._df) is ssql.DataFrame, "sourceDf must be a valid Pyspark dataframe"
 
-        return self.scriptDataGeneratorFromSchema(self._df.schema, suppressOutput=suppressOutput, name=name)
+        if self._dataSummary is None:
+            df_summary = self.summarizeToDF()
+
+            self._dataSummary = {}
+            for row in df_summary.collect():
+                row_key_pairs = row.asDict()
+                self._dataSummary[row['measure_']] = row_key_pairs
+
+        return self._scriptDataGeneratorCode(self._df.schema,
+                                             suppressOutput=suppressOutput,
+                                             name=name,
+                                             dataSummary=self._dataSummary,
+                                             sourceDf=self._df)