+++ tests/test_generation_from_data.py
@@ -46,7 +46,7 @@ class TestGenerationFromData:
             .withColumn("r_value", "float", expr="floor(rand() * 350) * (86400 + 3600)",
                         numColumns=(2, 4), structType="array")
             .withColumn("tf_flag", "boolean", expr="id % 2 = 1")
-            .withColumn("short_value", "short", max=32767)
+            .withColumn("short_value", "short", max=32767, percentNulls=0.1)
             .withColumn("byte_value", "tinyint", max=127)
             .withColumn("decimal_value", "decimal(10,2)", max=1000000)
             .withColumn("decimal_value", "decimal(10,2)", max=1000000)