+++ CHANGELOG.md
@@ -12,9 +12,9 @@ All notable changes to the Databricks Labs Data Generator will be documented in 
 * Additional build ordering enhancements to reduce circumstances where explicit base column must be specified
 
 #### Added
-* Added top level `random` attribute to data generator specification constructor
 * Scripting of data generation code from schema (Experimental)
 * Scripting of data generation code from dataframe (Experimental)
+* Added top level `random` attribute to data generator specification constructor
 
 
 ### Version 0.3.3post2